Some more changes from review, moved some tests from UsageStatisticsReportsSettingsDeterminerTest to ScenarioStatisticsTest as some functions where also moved to it's counterpart Todo: improve the tests as currently not every statistic calculation may be tested
