+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/statistics/ScenarioStatistics.scala
@@ -1,8 +1,11 @@
 package pl.touk.nussknacker.ui.statistics
 
-import cats.implicits.{toFoldableOps, toTraverseOps}
-import pl.touk.nussknacker.engine.api.component.ComponentType
+import cats.implicits.toFoldableOps
+import pl.touk.nussknacker.engine.api.component.{ComponentType, ProcessingMode}
+import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
+import pl.touk.nussknacker.engine.util.Implicits.RichScalaMap
 import pl.touk.nussknacker.restmodel.component
+import pl.touk.nussknacker.ui.process.processingtype.DeploymentManagerType
 import pl.touk.nussknacker.ui.process.repository.DbProcessActivityRepository
 
 import java.time.Instant
@@ -16,7 +19,7 @@ object ScenarioStatistics {
       val scenariosCount = scenariosInputData.count(!_.isFragment)
       //        Nodes stats
       val sortedNodes  = scenariosInputData.map(_.nodesCount).sorted
-      val nodesMedian  = sortedNodes.get(sortedNodes.length / 2).getOrElse(0)
+      val nodesMedian  = calculateMedian(sortedNodes)
       val nodesAverage = sortedNodes.sum / scenariosInputData.length
       val nodesMax     = sortedNodes.head
       val nodesMin     = sortedNodes.last
@@ -24,7 +27,7 @@ object ScenarioStatistics {
       val categoriesCount = scenariosInputData.map(_.scenarioCategory).toSet.size
       //        Version stats
       val sortedVersions  = scenariosInputData.map(_.scenarioVersion.value).sorted
-      val versionsMedian  = sortedVersions.get(sortedVersions.length / 2).getOrElse(0)
+      val versionsMedian  = calculateMedian(sortedVersions)
       val versionsAverage = sortedVersions.sum / scenariosInputData.length
       val versionsMax     = sortedVersions.head
       val versionsMin     = sortedVersions.last
@@ -32,10 +35,10 @@ object ScenarioStatistics {
       val authorsCount = scenariosInputData.map(_.createdBy).toSet.size
       //        Fragment stats
       val fragments        = scenariosInputData.map(_.fragmentsUsedCount).sorted
-      val fragmentsMedian  = fragments.get(fragments.length / 2).getOrElse(0)
+      val fragmentsMedian  = calculateMedian(fragments)
       val fragmentsAverage = fragments.sum / scenariosCount
       //          Uptime stats
-      val lastActions = scenariosInputData.map(_.lastDeployedAction).filter(_.isDefined).sequence.get
+      val lastActions = scenariosInputData.flatMap(_.lastDeployedAction)
       val sortedUptimes = lastActions.map { action =>
         Instant.now.getEpochSecond - action.performedAt.getEpochSecond
       }.sorted
@@ -102,12 +105,49 @@ object ScenarioStatistics {
       Map.empty
     } else {
       val withoutFragments      = componentList.filterNot(comp => comp.componentType == ComponentType.Fragment)
-      val componentsByNameCount = withoutFragments.groupBy(_.name)
+      val componentsByNameCount = withoutFragments.groupBy(_.name).size
 
       Map(
-        ComponentsCount -> componentsByNameCount.size
+        ComponentsCount -> componentsByNameCount
       ).map { case (k, v) => (k.toString, v.toString) }
     }
+  }
+
+  private val flinkDeploymentManagerType = DeploymentManagerType("flinkStreaming")
+
+  private val liteK8sDeploymentManagerType = DeploymentManagerType("lite-k8s")
+
+  private val liteEmbeddedDeploymentManagerType = DeploymentManagerType("lite-embedded")
+
+  private val knownDeploymentManagerTypes =
+    Set(flinkDeploymentManagerType, liteK8sDeploymentManagerType, liteEmbeddedDeploymentManagerType)
+
+  // We have four dimensions:
+  // - scenario / fragment
+  // - processing mode: streaming, r-r, batch
+  // - dm type: flink, k8s, embedded, custom
+  // - status: active (running), other
+  // We have two options:
+  // 1. To aggregate statistics for every combination - it give us 3*4*2 + 3*4 = 36 parameters
+  // 2. To aggregate statistics for every dimension separately - it gives us 2+3+4+1 = 10 parameters
+  // We decided to pick the 2nd option which gives a reasonable balance between amount of collected data and insights
+  private[statistics] def determineStatisticsForScenario(inputData: ScenarioStatisticsInputData): Map[String, Int] = {
+    Map(
+      "s_s"     -> !inputData.isFragment,
+      "s_f"     -> inputData.isFragment,
+      "s_pm_s"  -> (inputData.processingMode == ProcessingMode.UnboundedStream),
+      "s_pm_b"  -> (inputData.processingMode == ProcessingMode.BoundedStream),
+      "s_pm_rr" -> (inputData.processingMode == ProcessingMode.RequestResponse),
+      "s_dm_f"  -> (inputData.deploymentManagerType == flinkDeploymentManagerType),
+      "s_dm_l"  -> (inputData.deploymentManagerType == liteK8sDeploymentManagerType),
+      "s_dm_e"  -> (inputData.deploymentManagerType == liteEmbeddedDeploymentManagerType),
+      "s_dm_c"  -> !knownDeploymentManagerTypes.contains(inputData.deploymentManagerType),
+      "s_a"     -> inputData.status.contains(SimpleStateStatus.Running),
+    ).mapValuesNow(if (_) 1 else 0)
+  }
+
+  private def calculateMedian[T: Numeric](list: List[T]): T = {
+    list.get(list.size / 2).getOrElse(implicitly[Numeric[T]].zero)
   }
 
 }
